fixed int issue

+++ EastBarley/Models/PaymentTypes.cs
@@ -10,7 +10,7 @@ namespace EastBarley.Models
         public int PaymentId { get; set; }
         public int UserId { get; set; }
         public string PaymentType { get; set; }
-        public int AccountNumber { get; set; }
+        public long AccountNumber { get; set; }
         public int ExpirationYear { get; set; }
         public int ExpirationMonth { get; set; }
         public bool isActive { get; set; }
